+++ src/models/types/MetaComps.ts
@@ -1,0 +1,5 @@
+export interface MetaComps {
+    multihit: boolean;
+    admech: boolean;
+    neuro: boolean;
+}